# realtimechat-backend/src/main/java/com/mczapiewski/realtimechat/user/UserResponse.java
package com.mczapiewski.realtimechat.user;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {

    private String id;
    private String firstName;
    private String lastName;
    private String email;
    private LocalDateTime lastSeen;
    private boolean isOnline;
}